Move generic modals and views
