+++ helios/views.py
@@ -82,26 +82,3 @@ class VoiceView(discord.ui.View):
             )
 
 
-class YesNoView(discord.ui.View):
-    def __init__(self, author: discord.Member, *, timeout=5):
-        super().__init__(timeout=timeout)
-        self.author = author
-        self.value = None
-
-    @discord.ui.button(label='Yes', style=discord.ButtonStyle.green)
-    async def yes(self, interaction: discord.Interaction, button: discord.ui.Button):
-        if interaction.user == self.author:
-            self.value = True
-            await interaction.response.defer()
-            self.stop()
-        else:
-            await interaction.response.send_message('You are not allowed to respond', ephemeral=True)
-
-    @discord.ui.button(label='No', style=discord.ButtonStyle.red)
-    async def no(self, interaction: discord.Interaction, button: discord.ui.Button):
-        if interaction.user == self.author:
-            self.value = False
-            await interaction.response.defer()
-            self.stop()
-        else:
-            await interaction.response.send_message('You are not allowed to respond', ephemeral=True)
